Complete day 6

+++ 2023/src/day6.h
@@ -1,2 +1,2 @@
 int day6Part1();
-int day6Part2();
+long long day6Part2();
